1.12.7.13 Made gang value long (previously was int)

+++ modules/core/src/main/java/dev/drawethree/ultraprisoncore/gangs/commands/impl/value/GangValueAddSubCommand.java
@@ -33,7 +33,7 @@ public final class GangValueAddSubCommand extends GangSubCommand {
 					return false;
 				}
 
-				int amount = Integer.parseInt(args.get(1));
+				long amount = Long.parseLong(args.get(1));
 				String operation = "add";
 
 				return this.command.getPlugin().getGangsManager().modifyValue(sender, gang.get(), amount, operation);
